Handle realm_user add events. Add tests
